Add ntc-netbox-plugin-metrics-ext plugin
Through copying it to the -deploy directory and applying manually the
needed change (see PR bellow). Adapting freeze_requirements.sh
accordingly so it's included in the wheels.

Ideally only temporarily the time to check that it's working fine.
Longer term is either
https://github.com/networktocode/ntc-netbox-plugin-metrics-ext/pull/22
gets merged and a new pypi is released. Or we decide to maintain our own
fork.

Bug: T311052
Change-Id: I31a347a29d7b1e8c2e5d75eaaa0828890bbbf8af