Change build image user from root to nobody
The python-build images used to default to run as `root` which I have
changed to `nobody` in I03770e2fa7344a4b7c9b3ef9786e9a4a1bfa6cf8.
When building the image, switch to root in order to install Debian
packages and then switch back to nobody.

The freeze then run as nobody and pip attempts to save cache to the non
existent /nonexistent/.cache/pip directory. Create a cache directory on
the host, mount it in the container and instruct pip to use it via
XDG_CACHE_HOME.

Change the freeze script to install using a virtual environment.
Makefile invokes Docker as the host user and mount /etc/password, pip
would end up trying installing to a non existent homedir (eg:
/home/hashar).

Set UID to the Makefile invoking user, the error was not seen previously
since everything ran as root.

Add support for Podman by adding --userns=keep-id when it is detected.

Those changes mostly come from the integration/zuul/deploy repo.

Also refresh the image when Dockerfile.build.in is modified by making it
a prerequisite to the Dockerfile.build target.

Change-Id: I9cc9c17c206b262126cb6400c8bc17afd2ce0999